requirements.txt
numpy==1.26.4
pandas==2.2.1
pandas-ta==0.3.14b0
redis==5.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
